+++ .vscode/settings.json
@@ -10,6 +10,7 @@
         "string": "cpp",
         "unordered_map": "cpp",
         "vector": "cpp",
-        "system_error": "cpp"
+        "system_error": "cpp",
+        "classes_metodos_grafo.h": "c"
     }
 }
